+++ R/train.R
@@ -1786,11 +1786,13 @@ methods::setMethod(
     if(getOption('ibis.setupmessages', default = TRUE)) myLog('[Done]','green',paste0('Completed after ', round( as.numeric(out$settings$duration()), 2),' ',attr(out$settings$duration(),'units') ))
 
     # Quick check that the prediction is consistent with background extent
-    if(!is_comparable_raster(out$get_data(), x$background )){
+    if(settings$get('inference_only')==FALSE){
+      if(!is_comparable_raster(out$get_data(), x$background )){
       o <- out$get_data()
       o <- terra::extend(o, x$background) |> terra::crop(x$background)
       out <- out$set_data("prediction", o)
       try({ rm(o) })
+      }
     }
 
     # Clip to limits again to be sure